requirement.txt
# 核心依赖
streamlit>=1.28.0
requests>=2.31.0

# AI模型相关
google-generativeai>=0.3.0
huggingface-hub>=0.19.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 可视化（新增）
plotly>=5.17.0

# RAG系统（如果使用）
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# 可选：更强大的翻译支持
# deep-translator>=1.11.4
# googletrans==4.0.0rc1